clean up color for team creation to a new file
