+++ Wire-iOS/Sources/UserInterface/Registration/RegistrationTeam/LandingViewController.swift
@@ -20,18 +20,6 @@ import Foundation
 import UIKit
 import Cartography
 
-///FIXME: mv to new file
-fileprivate extension UIColor {
-    static let background = UIColor(red:0.97, green:0.97, blue:0.97, alpha:1.0)
-    static let inactiveButton = UIColor(red:0.20, green:0.22, blue:0.23, alpha:0.16)
-    static let activeButton = UIColor(red:0.14, green:0.57, blue:0.83, alpha:1.0)
-    static let createAccountBlue = UIColor(for: .strongBlue)!
-    static let createTeamGreen = UIColor(for: .strongLimeGreen)!
-    static let textColor = UIColor(red:0.20, green:0.22, blue:0.23, alpha:1.0)
-    static let subtitleColor = UIColor(red:0.20, green:0.22, blue:0.23, alpha:0.56)
-}
-
-
 @objc protocol LandingViewControllerDelegate {
     func landingViewControllerDidChooseCreateAccount()
     func landingViewControllerDidChooseCreateTeam()
